change floats to double

+++ src/xjadeo/xjadeo.c
@@ -109,9 +109,9 @@ uint64_t    osd_vtc_oob;
 //------------------------------------------------
 #ifdef WARP
 //osc scale modification
-float display_scale_x_modifier = 0;
-float display_scale_y_modifier = 0;
-float display_deform_corners[8]= {0};
+double display_scale_x_modifier = 0;
+double display_scale_y_modifier = 0;
+double display_deform_corners[8]= {0};
 int recalculate_homography = 1; // we need to calculte it the first time, then only recalculate when corresponding osc messages arrive
 #endif
